backend/pyproject.toml
[project]
name = "pixiv-download-helper-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "muesli", email = "lavenderfield385@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "piexif>=1.1.3",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruamel-yaml>=0.18.14",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
